change Task interface to type and rename the file
